requirements.txt
Faker
Flask
mysql-connector-python
flask-cors
